+++ src/main/resources/Create_Table.sql
@@ -5,22 +5,22 @@ CREATE TABLE `dragon_tiger_stock`
 	id					bigint primary key AUTO_INCREMENT,
     code   				varchar(20),
     name   				varchar(100),
-    pchange    			decimal(10,4),
-    amount       		decimal(15,2),
-    buy      			decimal(15,2),
-    sell                decimal(15,2),
+    pchange    			decimal(30,4),
+    amount       		decimal(30,2),
+    buy      			decimal(30,2),
+    sell                decimal(30,2),
     reason 				varchar(500),
-    bratio 				decimal(6,2),
-    sratio        		decimal(6,2),
+    bratio 				decimal(15,2),
+    sratio        		decimal(15,2),
     ts_date				date,
     broker				varchar(200),
-    count 				int,
-	probability       	decimal(6,2),
-	broker_buy   		decimal(15,2),
-	broker_buy_prop    	decimal(6,2),
-	broker_sell 		decimal(15,2),
-	broker_sell_prop    decimal(6,2),
-	net					decimal(15,2)
+    count 				bigint,
+	probability       	decimal(15,2),
+	broker_buy   		decimal(30,2),
+	broker_buy_prop    	decimal(15,2),
+	broker_sell 		decimal(30,2),
+	broker_sell_prop    decimal(15,2),
+	net					decimal(30,2)
 );
 
 ALTER TABLE `dragon_tiger_stock` ADD UNIQUE `unique_index`(`code`, `ts_date`, `broker`, `reason`);
@@ -30,11 +30,11 @@ CREATE TABLE `broker_buy_summary`
     id					bigint primary key AUTO_INCREMENT,
     broker_code   		varchar(20),
     broker_name      	varchar(100),
-    s_money    			decimal(10,4),
-    b_money       		decimal(15,2),
-    jm_money      		decimal(15,2),
-    b_count 		    int,
-    s_Count         	int,
+    s_money    			decimal(30,4),
+    b_money       		decimal(30,2),
+    jm_money      		decimal(30,2),
+    b_count 		    bigint,
+    s_Count         	bigint,
     ts_date				date
 );
